+++ source/bin.js
@@ -2,8 +2,7 @@
 
 // External
 import { resolve, join, dirname } from 'path'
-import e from 'errlop'
-const Errlop = e.default
+import Errlop from 'errlop'
 
 // get root with imports
 import url from 'url'
@@ -16,11 +15,11 @@ import boundation from './index.js'
 import state from './state.js'
 import { pwd } from './data.js'
 
-// Process unhandled rejections
-process.on('unhandledRejection', function unhandledRejection(error) {
-	console.error(new Errlop('An unhandled promise failed', error))
-	process.exit(-1)
-})
+// // Process unhandled rejections
+// process.on('unhandledRejection', function unhandledRejection(error) {
+// 	console.error(new Errlop('An unhandled promise failed', error))
+// 	process.exit(-1)
+// })
 
 async function main() {
 	// boundation